src/requirements.txt
pandas
numpy
matplotlib.pyplot
sklearn
seaborn
scipy
xgboost
folium
IPython.display